chore(dev.yml): update komune-io/fixers-gradle workflow to use a specific commit hash chore(docs.yml): remove unused docs workflow file
The dev.yml file was updated to use a specific commit hash (29b6c0d48bc37b8ee48f6128cfc6e67d19121932) of the komune-io/fixers-gradle workflow. This ensures that the workflow uses a specific version of the workflow file, providing stability and reproducibility.

The docs.yml file was removed as it was no longer needed and was unused. This helps to keep the repository clean and organized.