feat(dependencies): add Observability dependencies for micrometer, micrometerOtlp, and opentelemetry feat(settings): include f2-spring-boot-starter-observability-otel in project settings feat(sonar): enable summary comment for pull requests in SonarQube analysis
The changes introduce new Observability dependencies for micrometer, micrometerOtlp, and opentelemetry to enhance observability capabilities in the project. Additionally, the f2-spring-boot-starter-observability-otel module is included in the project settings to enable observability features. Lastly, the summary comment for pull requests is enabled in SonarQube analysis for better visibility and collaboration during code reviews.