feat(build): add support for OpenTelemetry properties configuration in F2 module
The Dependencies.kt file now includes a new function `configurationProcessor`
to add the "org.springframework.boot:spring-boot-configuration-processor"
dependency. This change allows for the configuration of OpenTelemetry
properties specific to the F2 module.

A new build.gradle.kts file is added for the f2-spring-boot-starter-observability-opentelemetry module. This file includes dependencies related to OpenTelemetry, such as micrometer and opentelemetry exporters.

New Kotlin files are added for defining OpenTelemetry properties specific to the F2 module, including F2OpenTelemetryProperties, F2OpenTelemetryMetricsProperties, and F2OpenTelemetryTracesProperties.

Additionally, a new OtelPropertiesListener.kt file is added to handle the application of OpenTelemetry properties based on the environment configuration.

Lastly, a new spring.factories file is added to register the OtelPropertiesListener as an ApplicationListener for the Spring application context.

These changes enable the configuration of OpenTelemetry properties specific to the F2 module, enhancing observability capabilities within the application.