feat(auth): add new module f2-spring-boot-starter-auth-tenant with dependencies and configuration files The changes introduce a new module called f2-spring-boot-starter-auth-tenant containing Kotlin files for AuthenticationProvider, PermissionEvaluator, and ConfigurationConstants. The build.gradle.kts file is added with necessary dependencies and plugins for the module. This new module enhances the authentication functionality of the application by providing classes for handling security context, authentication, authorization, and configuration constants.
feat(auth): add TrustedIssuerJwtAuthenticationManagerResolver class for resolving authentication managers based on trusted issuers
feat(auth): add TrustedIssuerJwtAuthenticationManagerResolver class to dynamically resolve ReactiveAuthenticationManagers based on trusted issuers. This class allows for building and caching authentication managers for specific issuers, improving the flexibility and security of the authentication process. Additionally, a new import for WebSecurityConfig is added to the AutoConfiguration imports file, and the f2-spring-boot-starter-auth-tenant module is included in the settings.gradle.kts file for future development.