feat(CoroutinesUtils.kt): add utility functions for coroutine support
Introduce CoroutinesUtils.kt to provide utility functions for handling
Kotlin coroutines in a Spring Cloud Function context. This includes
functions to determine types related to suspending functions, flows,
and continuations, as well as functions to invoke suspending functions,
suppliers, and consumers. These utilities enhance the ability to work
with asynchronous and reactive programming paradigms in Kotlin.

docs(function.http): update sample HTTP requests

Add new sample HTTP requests to the function.http file, including
requests for new endpoints `/uuid2` and `/uuids`. This helps in testing
and demonstrating the new functionalities provided by the application.

chore: remove sonar-project.properties

Delete the sonar-project.properties file as it is no longer needed.
This file was used for SonarCloud integration, which is not currently
required for the project. Removing it helps in cleaning up the project
directory and avoiding unnecessary configurations.